Adding initial orientational clustering and plotting capability for unit cell clustering
Orientational clustering is done using a Z-score metric. More work needed to add plotting capability. Plotting is only possible right now as a vs b (unit cell lengths).
For the sub-sampling portion, the random seed is currently set to a different number for each iteration in the loop otherwise the same set of spots are chosen repeatedly.