Returning a list of crystals determined through clustering results
Clusters are determined first through unit cell dimension. Datapoints in the same unit cell
cluster are further clustered based on their orientational matrix A. In the end, a list of such
crystal models are found and returned back by the consensus function. This reproduces the answers
in the case of the 2-image composite created from the simulated data. For each crystal model, we
reindex and then refine/integrate the data.