Adjusting detector distance to align calculated spots onto observed
The detector distance needs to be slightly modified for the experiments_centroids to best align all the calculated
spots on an image with the observed spots in the radial direction. This will ensure the prediction from the
indexing solution is as close as possible to the observed bragg spots. The adjustment is done
using dnew = d*((robs-r0)/(rcal-r0)) where r0 is the beam center, d and dnew are the detector distances.