Resetting detector distance after indexing prior to refinement.
Note that currently outlier rejection is not turned off during the refinement step
after detector distance is reset. Refinement seems to complete OK on the 100 test images.
This may need to be looked at in future.
Also adding a script to compare dR = |rcal-robs| for the analysis in the test image in forNKS